explicit casts

+++ benchmark/src/libfork_benchmark/fib/serial_return.cpp
@@ -21,7 +21,7 @@ void fib_serial_return(benchmark::State &state) {
   std::int64_t n = state.range(0);
   std::int64_t expect = fib_ref(n);
 
-  state.counters["n"] = n;
+  state.counters["n"] = static_cast<double>(n);
 
   for (auto _ : state) {
     benchmark::DoNotOptimize(n);
